+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Better error message when the device is low on storage
+* Fixed a bug where Gallery selection of document images did not work in some cases
 
 ## 10.1.1